+++ backend/src/main/java/com/yuuka/backend/sinkingfund/infrastructure/SinkingFundBalanceProjection.java
@@ -1,11 +1,12 @@
 package com.yuuka.backend.sinkingfund.infrastructure;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface SinkingFundBalanceProjection {
   UUID getSinkingFundId();
 
-  Long getCurrentBalanceMinor();
+  BigDecimal getCurrentBalanceMinor();
 
   Long getTransactionCount();
 }
